Configurado Upload de arquivos public e private para AWS
Close #19
